requirements.txt
youtube_transcript_api
diffusers
torch
transformers
requests
openai
python-dotenv
markdown